Update plenario worker tasks with celery
