+++ manage.py
@@ -6,7 +6,7 @@ from flask_script import Manager
 from os import getenv
 
 from plenario import create_app as server
-from plenario.update import create_worker as worker
+from plenario.worker import create_worker as worker
 
 
 apps = {
